--- packages/ui-toolkit/src/components/atoms/Button_V1/ButtonV1.constants.ts
@@ -1,19 +1,0 @@
-export const VARIANTS = {
-  PRIMARY: 'Primary',
-  SECONDARY: 'Secondary',
-  TERTIARY: 'Tertiary',
-  POSITIVE: 'Positive',
-  NEGATIVE: 'Negative'
-};
-
-
-export const BUTTON_SIZES = {
-  SMALL: 'Small',
-  BASE: 'Base',
-  LARGE: 'Large'
-};
-
-export const ROLE = {
-  BUTTON: 'button',
-  LINK: 'link'
-};
