--- packages/ui-toolkit/src/components/atoms/Button_V1/buttonV1.css
@@ -1,149 +1,0 @@
-.mint-btn-default {
-  border-radius: 8px;
-  flex-shrink: 0;
-  border: 0;
-  outline: none;
-  min-width: 72px;
-}
-
-.mint-btn-default:hover {
-  opacity: 0.96;
-}
-
-.mint-btn-default:focus-visible {
-  outline-offset: 2px;
-  outline: 2px solid -webkit-focus-ring-color; 
-  /* // default color */
-}
-
-.mint-btn-default:active {
-  opacity: 0.96;
-}
-
-.mint-btn-cursor-default {
-  cursor: default;
-}
-
-.mint-btn-border-no-accent{
-  border: 1px solid var(--gray150) !important;
-}
-
-/* -- Size Classes -- */
-
-.mint-btn-small {
-  padding: 12px;
-  height: 32px;
-  gap: 6px;
-  width: fit-content;
-}
-
-.mint-btn-medium {
-  padding: 12px 16px;
-  height: 40px;
-  width: fit-content;
-  gap: 6px;
-}
-
-.mint-btn-large {
-  padding: 16px 24px;
-  width: fit-content;
-  height: 48px;
-  gap: 8px;
-}
-
-.mint-btn-full-width{
-  width: 100%;
-}
-
-/* -- Disabled Button -- */
-
-.mint-btn-disabled{
-  cursor: default;
-  pointer-events: none;
-  background-color: var(--gray150);
-}
-
-.mint-btn-disabled.mint-btn-tertiary-disabled{
-  background-color: transparent;
-}
-/* -- Seconday Button -- */
-
-.mint-btn-secondary-no-accent:hover {
-  background-color: var(--gray50);
-}
-
-/* -- Tertairy Button -- */
-
-/* when isAccent = true */
-.mint-btn-tertiary-accent:hover{
-  background-color: var(--green100);
-}
-
-/* when isAccent = false */
-.mint-btn-tertiary-no-accent:hover{
-  background-color: var(--gray50);
-}
-
-.mint-btn-tertiary-border {
-  border-style: dashed !important;
-  border-width: 0 0 1px 0 !important;
-}
-
-/* -- Compact classes -- */
-
-.mint-btn-compact {
-  width: fit-content;
-  min-width: fit-content;
-  height: fit-content;
-  border-radius: 0;
-  padding: 0 !important;
-}
-
-
-/*  -- Loader Button --  */
-
-.mint-btn-loader {
-  color: transparent !important;
-}
-
-.mint-btn-loader-container {
-  position: absolute;
-  margin-top: 4px;
-}
-
-.mint-btn-loader-primary {
-  border: 2px solid var(--white); /* the loader color will always be white in dark as well as in light theme */
-  border-top-color: transparent;
-}
-
-.mint-btn-loader-no-accent { /*used because the token doesn't override the existing color of loader*/
-  border: 2px solid var(--gray900);
-  border-right-color: transparent;
-}
-
-.mint-btn-small .mint-btn-loader-item {
-  width: 16px;
-  height: 16px;
-}
-
-.mint-btn-medium .mint-btn-loader-item {
-  width: 20px; 
-  height: 20px;
-}
-
-.mint-btn-large .mint-btn-loader-item {
-  width: 24px;
-  height: 24px;
-}
-
-/*  -- Fixed Bottom -- */
-
-.mint-btn-fixed-bottom {
-  position: fixed;
-  left: 0;
-  border-width: 1px 0 0 0 !important;
-  padding: 12px;
-  width: 100%;
-  z-index: 1000;
-  bottom: 0;
-}
