--- packages/ui-toolkit/src/components/atoms/Button_V1/index.ts
@@ -1,2 +1,0 @@
-export { default as ButtonV1 } from './ButtonV1';
-export { VARIANTS, BUTTON_SIZES } from './ButtonV1.constants';
